chore: update header

+++ interfaces/json/api_dtos.py
@@ -53,7 +53,7 @@ class DependencyDetail(BaseModel):
     primary_lvl_years: Optional[int] = None
     secondary_lvll_years: Optional[int] = None
     tertiary_lvll_years: Optional[int] = None
-    tuition_fee_incr_perc: Optional[int] = None
+    tuition_fee_incr_perc: Optional[float] = None
     dependency_provision: Optional[DependencyProvision] = None
     
 class DependencyDetailPostJson(BaseModel):
